+++ lib/rubocop.rb
@@ -139,6 +139,7 @@ require 'rubocop/cop/performance/redundant_sort_by'
 require 'rubocop/cop/performance/reverse_each'
 require 'rubocop/cop/performance/sample'
 require 'rubocop/cop/performance/size'
+require 'rubocop/cop/performance/start_with'
 require 'rubocop/cop/performance/string_replacement'
 
 require 'rubocop/cop/style/access_modifier_indentation'